+++ src/index.php
@@ -105,6 +105,12 @@
     */
    define('DIR_SYS', DIR_INDEX . 'sys' . DIRECTORY_SEPARATOR);
 
+   /**
+    * The minimum available PHP integer
+    * @var int
+    */
+   define('PHP_INT_MIN',~PHP_INT_MAX);
+
    // Change dir to make sure CLI requests are correct
    if (defined('STDIN')) {
       chdir(dirname(__FILE__));