+++ changelog.md
@@ -17,6 +17,7 @@ Added features
 * clearCrontab() method at crontab editor
 * Possibility to automatically commit crontab changes. Use with caution!
 * getAtIndex() method in Cron
+* \Alo\Statics\Security class added
 
 Miscellaneous
 
@@ -26,6 +27,7 @@ Miscellaneous
 * FormValidator documentation edit
 * Cron editor now throws an exception if you try to invoke it on a Windows machine
 * PHPUNIT_RUNNING constant definition moved out of source code. It now simply checks if the constant is defined in the bootstrap file.
+* The global functions getFingerprint(), getUniqid(), escape() moved to \Alo\Statics\Security. The escape() function has been renamed to un_xss() in the class.
 
 # 0.2.1 (2015-05-05) #
 Major bugfixes